Move nested types to own files
